refactor(sandbox): restructure agent sandbox runtime and preparation lifecycle
