+++ packages/global/core/ai/sandbox/type.ts
@@ -1,0 +1,7 @@
+import z from 'zod';
+
+export const SandboxImageConfigSchema = z.object({
+  repository: z.string(),
+  tag: z.string().optional()
+});
+export type SandboxImageConfigType = z.infer<typeof SandboxImageConfigSchema>;
