+++ packages/service/core/ai/sandbox/runtime/profile/opensandbox.ts
@@ -9,7 +9,6 @@ import {
 import { OPEN_SANDBOX_DEFAULT_ROOT_PATH } from '@fastgpt-sdk/sandbox-adapter';
 
 const OPEN_SANDBOX_ENTRYPOINT = '/home/sandbox/entrypoint.sh';
-const OPEN_SANDBOX_HOME_DIRECTORY = '/home/sandbox';
 const OPEN_SANDBOX_DOCKER_LOCAL_NETWORK_POLICY = {
   defaultAction: 'allow' as const,
   egress: [
@@ -40,7 +39,6 @@ export function buildOpenSandboxRuntimeProfile(): SandboxRuntimeProfile {
     provider: 'opensandbox',
     defaultImage,
     workDirectory,
-    homeDirectory: OPEN_SANDBOX_HOME_DIRECTORY,
     entrypoint: OPEN_SANDBOX_ENTRYPOINT,
     skillsRootPath: getSandboxSkillsRootPath(workDirectory),
     buildConfig(input = {}) {
