+++ packages/service/core/ai/sandbox/runtime/profile/sealosdevbox.ts
@@ -1,10 +1,7 @@
 import { serviceEnv } from '../../../../../env';
-import { SandboxTypeEnum } from '@fastgpt/global/core/ai/skill/constants';
 import type { SandboxRuntimeProfile } from './types';
 import { getSandboxSkillsRootPath, mergeStringRecord, mergeUnknownRecord } from './utils';
 import { parseImageSpec } from '@fastgpt-sdk/sandbox-adapter';
-
-const SEALOS_DEVBOX_HOME_DIRECTORY = '/home/devbox';
 
 /**
  * 构建 Sealos Devbox 的 FastGPT 运行态 profile。
@@ -20,7 +17,6 @@ export function buildSealosRuntimeProfile(): SandboxRuntimeProfile {
     provider: 'sealosdevbox',
     defaultImage,
     workDirectory,
-    homeDirectory: SEALOS_DEVBOX_HOME_DIRECTORY,
     entrypoint: '',
     skillsRootPath: getSandboxSkillsRootPath(workDirectory),
     buildConfig(input = {}) {
