+++ packages/service/core/ai/sandbox/runtime/profile/utils.ts
@@ -1,10 +1,4 @@
-/** 去掉 sandbox 路径右侧斜杠，根路径保持可继续拼接的空前缀。 */
-export const trimSandboxPathRight = (value: string) =>
-  value === '/' ? '' : value.replace(/\/+$/, '');
-
-/** 用 sandbox 语义拼接路径，避免不同 provider 工作目录末尾斜杠导致双斜杠。 */
-export const joinSandboxPath = (basePath: string, path: string) =>
-  `${trimSandboxPathRight(basePath)}/${path}`;
+import { joinSandboxPath } from '../utils';
 
 /** FastGPT 约定所有 skill 包都写入运行态工作目录下的 skills 子目录。 */
 export const getSandboxSkillsRootPath = (workDirectory: string) =>
