+++ packages/service/core/ai/skill/package/constants.ts
@@ -12,9 +12,6 @@ node_modules/
 .pnpm-store/
 package.zip
 
-# FastGPT internal runtime files
-.fastgpt/
-
 # Python dependencies & cache
 __pycache__/
 *.pyc
