+++ packages/service/core/workflow/dispatch/ai/agent/index.ts
@@ -33,11 +33,7 @@ import {
 import { i18nT } from '@fastgpt/global/common/i18n/utils';
 import { getErrText } from '@fastgpt/global/common/error/utils';
 import type { InteractiveNodeResponseType } from '@fastgpt/global/core/workflow/template/system/interactive/type';
-import {
-  agentSandboxBootstrap,
-  ensureAgentSandboxRuntime,
-  streamAgentSandboxInitStatus
-} from './sub/sandbox';
+import { ensureAgentSandboxRuntime, streamAgentSandboxInitStatus } from './sub/sandbox';
 import type { WorkflowNodeResponseWriter } from '../../../../chat/nodeResponseStorage';
 import type { RuntimeNodeResponseSummary } from '../../type';
 import { createAgentNodeResponseCollector } from './nodeResponseCollector';
@@ -223,7 +219,6 @@ export const dispatchRunAgent = async (props: DispatchAgentModuleProps): Promise
       teamId: runningAppInfo.teamId,
       tmbId: runningUserInfo.tmbId,
       needSandboxRuntime: effectiveUseAgentSandbox,
-      sandboxBootstrap: agentSandboxBootstrap,
       sandboxEntrypoint: effectiveSandboxEntrypoint,
       skillIds,
       editSkillId,
