+++ packages/service/test/core/ai/sandbox/runtime/profile.test.ts
@@ -44,7 +44,6 @@ describe('sandbox runtime profile', () => {
         tag: 'stable'
       },
       workDirectory: '/workspace',
-      homeDirectory: '/home/sandbox',
       entrypoint: '/home/sandbox/entrypoint.sh'
     });
     expect(runtimeProfile.skillsRootPath).toBe('/workspace/skills');
@@ -59,7 +58,6 @@ describe('sandbox runtime profile', () => {
     expect(runtimeProfile).toMatchObject({
       provider: 'sealosdevbox',
       workDirectory: '/home/devbox/workspace',
-      homeDirectory: '/home/devbox',
       entrypoint: ''
     });
     expect(runtimeProfile.skillsRootPath).toBe('/home/devbox/workspace/skills');
